# Optimize advice with FilteredAttributes