--- a/api/all/src/main/java/io/opentelemetry/api/internal/ImmutableKeyValuePairs.java
+++ b/api/all/src/main/java/io/opentelemetry/api/internal/ImmutableKeyValuePairs.java
@@ -275,6 +275,10 @@ public abstract class ImmutableKeyValuePairs<K, V> {
     return sb.toString();
   }
 
+  /**
+   * Return the backing data array for these attributes. This is only exposed for internal use by
+   * opentelemetry authors. The contents of the array MUST NOT be modified.
+   */
   @SuppressWarnings("AvoidObjectArrays")
   public Object[] getData() {
     return data;