--- a/sdk/metrics/src/jmh/java/io/opentelemetry/sdk/metrics/MetricAdviceBenchmark.java
+++ b/sdk/metrics/src/jmh/java/io/opentelemetry/sdk/metrics/MetricAdviceBenchmark.java
@@ -11,7 +11,7 @@ import io.opentelemetry.api.incubator.metrics.ExtendedLongCounterBuilder;
 import io.opentelemetry.api.metrics.LongCounter;
 import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.sdk.testing.exporter.InMemoryMetricReader;
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.openjdk.jmh.annotations.Benchmark;
@@ -36,20 +36,68 @@ import org.openjdk.jmh.annotations.Warmup;
 @Fork(1)
 public class MetricAdviceBenchmark {
 
-  private static final Attributes ALL_ATTRIBUTES;
-  private static final Attributes SOME_ATTRIBUTES;
-  private static final List<AttributeKey<?>> SOME_ATTRIBUTE_KEYS;
+  static final AttributeKey<String> HTTP_REQUEST_METHOD =
+      AttributeKey.stringKey("http.request.method");
+  static final AttributeKey<String> URL_PATH = AttributeKey.stringKey("url.path");
+  static final AttributeKey<String> URL_SCHEME = AttributeKey.stringKey("url.scheme");
+  static final AttributeKey<Long> HTTP_RESPONSE_STATUS_CODE =
+      AttributeKey.longKey("http.response.status_code");
+  static final AttributeKey<String> HTTP_ROUTE = AttributeKey.stringKey("http.route");
+  static final AttributeKey<String> NETWORK_PROTOCOL_NAME =
+      AttributeKey.stringKey("network.protocol.name");
+  static final AttributeKey<Long> SERVER_PORT = AttributeKey.longKey("server.port");
+  static final AttributeKey<String> URL_QUERY = AttributeKey.stringKey("url.query");
+  static final AttributeKey<String> CLIENT_ADDRESS = AttributeKey.stringKey("client.address");
+  static final AttributeKey<String> NETWORK_PEER_ADDRESS =
+      AttributeKey.stringKey("network.peer.address");
+  static final AttributeKey<Long> NETWORK_PEER_PORT = AttributeKey.longKey("network.peer.port");
+  static final AttributeKey<String> NETWORK_PROTOCOL_VERSION =
+      AttributeKey.stringKey("network.protocol.version");
+  static final AttributeKey<String> SERVER_ADDRESS = AttributeKey.stringKey("server.address");
+  static final AttributeKey<String> USER_AGENT_ORIGINAL =
+      AttributeKey.stringKey("user_agent.original");
 
-  static {
-    SOME_ATTRIBUTES =
-        Attributes.builder()
-            .put("http.request.method", "GET")
-            .put("http.route", "/v1/users/{id}")
-            .put("http.response.status_code", 200)
-            .build();
-    ALL_ATTRIBUTES =
-        SOME_ATTRIBUTES.toBuilder().put("http.url", "http://localhost:8080/v1/users/123").build();
-    SOME_ATTRIBUTE_KEYS = new ArrayList<>(SOME_ATTRIBUTES.asMap().keySet());
+  static final List<AttributeKey<?>> httpServerMetricAttributeKeys =
+      Arrays.asList(
+          HTTP_REQUEST_METHOD,
+          URL_SCHEME,
+          HTTP_RESPONSE_STATUS_CODE,
+          HTTP_ROUTE,
+          NETWORK_PROTOCOL_NAME,
+          SERVER_PORT,
+          NETWORK_PROTOCOL_VERSION,
+          SERVER_ADDRESS);
+
+  static Attributes httpServerMetricAttributes() {
+    return Attributes.builder()
+        .put(HTTP_REQUEST_METHOD, "GET")
+        .put(URL_SCHEME, "http")
+        .put(HTTP_RESPONSE_STATUS_CODE, 200)
+        .put(HTTP_ROUTE, "/v1/users/{id}")
+        .put(NETWORK_PROTOCOL_NAME, "http")
+        .put(SERVER_PORT, 8080)
+        .put(NETWORK_PROTOCOL_VERSION, "1.1")
+        .put(SERVER_ADDRESS, "localhost")
+        .build();
+  }
+
+  static Attributes httpServerSpanAttributes() {
+    return Attributes.builder()
+        .put(HTTP_REQUEST_METHOD, "GET")
+        .put(URL_PATH, "/v1/users/123")
+        .put(URL_SCHEME, "http")
+        .put(HTTP_RESPONSE_STATUS_CODE, 200)
+        .put(HTTP_ROUTE, "/v1/users/{id}")
+        .put(NETWORK_PROTOCOL_NAME, "http")
+        .put(SERVER_PORT, 8080)
+        .put(URL_QUERY, "with=email")
+        .put(CLIENT_ADDRESS, "192.168.0.17")
+        .put(NETWORK_PEER_ADDRESS, "192.168.0.17")
+        .put(NETWORK_PEER_PORT, 11265)
+        .put(NETWORK_PROTOCOL_VERSION, "1.1")
+        .put(SERVER_ADDRESS, "localhost")
+        .put(USER_AGENT_ORIGINAL, "okhttp/1.27.2")
+        .build();
   }
 
   @State(Scope.Benchmark)
@@ -83,7 +131,7 @@ public class MetricAdviceBenchmark {
 
   @SuppressWarnings("ImmutableEnumChecker")
   public enum InstrumentParam {
-    NO_ADVICE_RECORD_ALL(
+    NO_ADVICE_ALL_ATTRIBUTES(
         new Instrument() {
           private LongCounter counter;
 
@@ -94,10 +142,24 @@ public class MetricAdviceBenchmark {
 
           @Override
           void record(long value) {
-            counter.add(value, ALL_ATTRIBUTES);
+            counter.add(value, httpServerSpanAttributes());
           }
         }),
-    ADVICE_RECORD_ALL(
+    NO_ADVICE_FILTERED_ATTRIBUTES(
+        new Instrument() {
+          private LongCounter counter;
+
+          @Override
+          void setup(Meter meter) {
+            counter = ((ExtendedLongCounterBuilder) meter.counterBuilder("counter")).build();
+          }
+
+          @Override
+          void record(long value) {
+            counter.add(value, httpServerMetricAttributes());
+          }
+        }),
+    ADVICE_ALL_ATTRIBUTES(
         new Instrument() {
           private LongCounter counter;
 
@@ -105,16 +167,16 @@ public class MetricAdviceBenchmark {
           void setup(Meter meter) {
             counter =
                 ((ExtendedLongCounterBuilder) meter.counterBuilder("counter"))
-                    .setAttributesAdvice(SOME_ATTRIBUTE_KEYS)
+                    .setAttributesAdvice(httpServerMetricAttributeKeys)
                     .build();
           }
 
           @Override
           void record(long value) {
-            counter.add(value, ALL_ATTRIBUTES);
+            counter.add(value, httpServerSpanAttributes());
           }
         }),
-    ADVICE_RECORD_SOME(
+    ADVICE_FILTERED_ATTRIBUTES(
         new Instrument() {
           private LongCounter counter;
 
@@ -122,13 +184,13 @@ public class MetricAdviceBenchmark {
           void setup(Meter meter) {
             counter =
                 ((ExtendedLongCounterBuilder) meter.counterBuilder("counter"))
-                    .setAttributesAdvice(SOME_ATTRIBUTE_KEYS)
+                    .setAttributesAdvice(httpServerMetricAttributeKeys)
                     .build();
           }
 
           @Override
           void record(long value) {
-            counter.add(value, SOME_ATTRIBUTES);
+            counter.add(value, httpServerMetricAttributes());
           }
         });