During the 8/7/2024 APAC java SIG, we discussed options for improving agent performance, including the impact of the `AdviceAttributesProcessor` (https://github.com/open-telemetry/opentelemetry-java/issues/6601).

This PR improve performance by wrapping Attributes implementation in a FilteredAttributes. FilteredAttributes uses the same memory allocation for key / values as its backing Attributes, but also retains a list of indices which have been filtered out. FilteredAttributes implements the Attributes contract with logic similar to the backing ImmutableKeyValueAttributes, but incorporating knowledge of indexes which should be filtered.

The performance improvement is quite good.

This PR adds MetricAdviceBenchmark with a variety of scenarios illustrating different interesting aspects of advice. The scenario of interest is `ADVICE_ALL_ATTRIBUTES`, which records the full set of HTTP server span attributes and applies attribute advice to narrow them to the HTTP server metric attributes. This is meant to be representative of a typical HTTP server instrumentation in `opentelemetry-java-instrumentation`.

The performance improvement is quite good: 

Before (baseline):
```
Benchmark                                                      (instrumentParam)  Mode  Cnt     Score     Error   Units
MetricAdviceBenchmark.record                               ADVICE_ALL_ATTRIBUTES  avgt   10   954.733 ±  14.308   ns/op
MetricAdviceBenchmark.record:gc.alloc.rate                 ADVICE_ALL_ATTRIBUTES  avgt   10  1445.940 ±  21.237  MB/sec
MetricAdviceBenchmark.record:gc.alloc.rate.norm            ADVICE_ALL_ATTRIBUTES  avgt   10  1447.989 ±   0.133    B/op
MetricAdviceBenchmark.record:gc.count                      ADVICE_ALL_ATTRIBUTES  avgt   10    24.000            counts
MetricAdviceBenchmark.record:gc.time                       ADVICE_ALL_ATTRIBUTES  avgt   10    15.000                ms
```

After:
```
Benchmark                                                      (instrumentParam)  Mode  Cnt     Score     Error   Units
MetricAdviceBenchmark.record                               ADVICE_ALL_ATTRIBUTES  avgt   10   624.285 ±  11.537   ns/op
MetricAdviceBenchmark.record:gc.alloc.rate                 ADVICE_ALL_ATTRIBUTES  avgt   10  1246.238 ±  22.510  MB/sec
MetricAdviceBenchmark.record:gc.alloc.rate.norm            ADVICE_ALL_ATTRIBUTES  avgt   10   816.010 ±   0.022    B/op
MetricAdviceBenchmark.record:gc.count                      ADVICE_ALL_ATTRIBUTES  avgt   10    21.000            counts
MetricAdviceBenchmark.record:gc.time                       ADVICE_ALL_ATTRIBUTES  avgt   10    16.000                ms
```

- CPU down from from 954 ns/op to 624 ns/op
- Memory allocation down from 1447 B/op to 816 B/op
- Of the 816 bytes allocated, 784 are the baseline needed to store the HTTP server span attributes. We're able to create a filtered view of the span attributes with just 816-784=32 bytes and 125ns of CPU time!
- Creating a copy of the metric attributes (i.e. what we do today) requires 512 bytes and 405 ns. So we've managed a 1-(32/512)=~94% reduction in memory, and 1-(125/405)=~70% CPU time reduction. Nice!

cc @wgy035, @open-telemetry/java-instrumentation-maintainers